+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-06-17 23:08:55
-// Commit: fix(state): fix memory leak by cancelling state subscription
+// Updated at: 2026-06-18 09:56:05
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '2096';
-  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
-  final DateTime timestamp = DateTime.parse('2026-06-17T23:08:55+05:00');
+  final String commitRef = '2097';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-06-18T09:56:05+05:00');
 }